Fixing pmt_calibration code
