+++ Run3Detector/analysis/pmt_calibration/peak_templates/#38138314735b6be765fa577249eace5919c38811#
@@ -1,0 +1,45 @@
+#!/usr/bin/env python3
+import ROOT as r
+import sys
+from array import array
+from pmt-calibration.python.TemplateTools import averageOfLists
+# Load in data file
+dataDir = "/home/rsantos/Data/PostProcessData/"
+peakTemplateFile = r.TFile(dataDir + sys.argv[1])
+tree = peakTemplateFile.Get("Events")
+
+# Define area to store plots
+if len(sys.argv) == 3:
+    plotDir = sys.argv[2]
+else:
+    plotDir = "."
+
+# Plot example waveform and save it in a ROOT file
+tree.GetEvent(1)
+voltages = array('d', tree.voltages)
+times = array('d', tree.times)
+
+# Average together all peaks and select time window to include laser-induced signal
+print(f"There are {tree.GetEntries()} entries in this ROOT file")
+voltageList = []
+for event in tree:
+    voltageList.append(tree.voltages)
+
+voltageAverage = averageOfLists(voltageList)
+
+r.TGraph(len(voltageAverage), array(voltageAverage))
+
+
+
+
+
+
+# Determine Noise
+
+
+
+# Determine Voltage Range
+
+
+#graph = r.TGraph(len(voltages), times, voltages)
+#graph.Draw()
